backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
python-multipart==0.0.5
aiofiles==0.7.0
python-docx==0.8.11
PyPDF2==2.0.0
google-cloud-aiplatform==1.25.0
python-dotenv==0.19.0
SQLAlchemy==1.4.23